+++ egui_demo_lib/src/apps/demo/demo_windows.rs
@@ -86,7 +86,7 @@ impl DemoWindows {
 
                 ui.separator();
 
-                if ui.button("Organize windows").clicked {
+                if ui.button("Organize windows").clicked() {
                     ui.ctx().memory().reset_areas();
                 }
             });
@@ -255,13 +255,13 @@ fn show_menu_bar(ui: &mut Ui) {
 
     menu::bar(ui, |ui| {
         menu::menu(ui, "File", |ui| {
-            if ui.button("Organize windows").clicked {
+            if ui.button("Organize windows").clicked() {
                 ui.ctx().memory().reset_areas();
             }
             if ui
                 .button("Clear egui memory")
                 .on_hover_text("Forget scroll, collapsing headers etc")
-                .clicked
+                .clicked()
             {
                 *ui.ctx().memory() = Default::default();
             }